requirements.txt
twitch.py==3.2.6
python-dotenv==1.0.1
langchain-ollama==0.2.2